Refactor Unkown Entity 15
Pulls out `EntityUnkId15` and maps related symbols.
