+++ config/symbols.pspeu.stwrp.txt
@@ -79,8 +79,8 @@ c_GoldPrizes = 0x092464D0;
 c_HeartPrizes = 0x09246500;
 D_80180EC4 = 0x09246508;
 D_80180F70 = 0x092465C8;
-D_80180FF8 = 0x09246690;
-D_80181008 = 0x092466A0;
+unk15_rot = 0x09246690;
+unk15_yVel = 0x092466A0;
 unk14_yVel = 0x092466B8;
 unk14_startFrame = 0x092466D0;
 unk14_lifetime = 0x092466D8;
