+++ config/symbols.stmad.txt
@@ -27,6 +27,8 @@ g_eUnk14SpawnRots = 0x80180FAC;
 unk14_yVel = 0x80180FE4;
 unk14_startFrame = 0x80180FFC;
 unk14_lifetime = 0x80181000;
+unk15_rot = 0x80180FBC;
+unk15_yVel = 0x80180FCC;
 g_Rooms = 0x80181130;
 g_eRedDoorUV = 0x801809D4;
 g_eRedDoorTiles = 0x80180594;
