+++ config/symbols.us.stcen.txt
@@ -24,6 +24,8 @@ c_GoldPrizes = 0x80180F58;
 c_HeartPrizes = 0x80180FE8;
 g_bigRedFireballAnim = 0x80181014;
 g_eUnk14SpawnRots = 0x80181110;
+unk15_rot = 0x80181120;
+unk15_yVel = 0x80181130;
 unk14_yVel = 0x80181148;
 unk14_startFrame = 0x80181160;
 unk14_lifetime = 0x80181164;
