+++ config/symbols.us.stdre.txt
@@ -31,6 +31,8 @@ c_GoldPrizes = 0x8018111C;
 c_HeartPrizes = 0x801811AC;
 g_bigRedFireballAnim = 0x801811D8;
 g_eUnk14SpawnRots = 0x801812D4;
+unk15_rot = 0x801812E4;
+unk15_yVel = 0x801812F4;
 unk14_yVel = 0x8018130C;
 unk14_startFrame = 0x80181324;
 unk14_lifetime = 0x80181328;
