+++ config/symbols.us.stno3.txt
@@ -31,6 +31,8 @@ c_GoldPrizes = 0x80182460;
 c_HeartPrizes = 0x801824F0;
 g_bigRedFireballAnim = 0x8018251C;
 g_eUnk14SpawnRots = 0x80182618;
+unk15_rot = 0x80182628;
+unk15_yVel = 0x80182638;
 unk14_yVel = 0x80182650;
 unk14_startFrame = 0x80182668;
 unk14_lifetime = 0x8018266C;
