+++ config/symbols.us.stnp3.txt
@@ -31,6 +31,8 @@ c_GoldPrizes = 0x80181DEC;
 c_HeartPrizes = 0x80181E7C;
 g_bigRedFireballAnim = 0x80181EA8;
 g_eUnk14SpawnRots = 0x80181FA4;
+unk15_rot = 0x80181FB4;
+unk15_yVel = 0x80181FC4;
 unk14_yVel = 0x80181FDC;
 unk14_startFrame = 0x80181FF4;
 unk14_lifetime = 0x80181FF8;
