+++ config/symbols.us.stnz0.txt
@@ -31,6 +31,8 @@ c_GoldPrizes = 0x80181D14;
 c_HeartPrizes = 0x80181DA4;
 g_bigRedFireballAnim = 0x80181DD0;
 g_eUnk14SpawnRots = 0x80181ECC;
+unk15_rot = 0x80181EDC;
+unk15_yVel = 0x80181EEC;
 unk14_yVel = 0x80181F04;
 unk14_startFrame = 0x80181F1C;
 unk14_lifetime = 0x80181F20;
