+++ config/symbols.us.strwrp.txt
@@ -19,6 +19,8 @@ c_GoldPrizes = 0x80180E30;
 c_HeartPrizes = 0x80180EC0;
 g_bigRedFireballAnim = 0x80180EEC;
 g_eUnk14SpawnRots = 0x80180FE8;
+unk15_rot = 0x80180FF8;
+unk15_yVel = 0x80181008;
 unk14_yVel = 0x80181020;
 unk14_startFrame = 0x80181038;
 unk14_lifetime = 0x8018103C;
