+++ config/symbols.us.stst0.txt
@@ -20,6 +20,8 @@ g_eRedDoorUV = 0x80181978;
 c_GoldPrizes = 0x80181CF0;
 g_bigRedFireballAnim = 0x80181DA4;
 g_eUnk14SpawnRots = 0x80181EA0;
+unk15_rot = 0x80181EB0;
+unk15_yVel = 0x80181EC0;
 unk14_yVel = 0x80181ED8;
 unk14_startFrame = 0x80181EF0;
 unk14_lifetime = 0x80181EF4;
