+++ src/st/cen/18084.c
@@ -36,7 +36,7 @@ void EntityUnkId13(Entity* entity) {
 
 #include "../entity_unkId14.h"
 
-INCLUDE_ASM("st/cen/nonmatchings/18084", EntityUnkId15);
+#include "../entity_unkId15.h"
 
 #include "../entity_olrox_drool.h"
