+++ src/st/mad/15520.c
@@ -85,35 +85,7 @@ void func_8019572C(Entity* entity) {
 
 #include "../entity_unkId14.h"
 
-void EntityUnkId15(Entity* entity) {
-    u16 temp_v0;
-
-    if (entity->step == 0) {
-        entity->flags = FLAG_UNK_2000 | FLAG_UNK_04000000 | FLAG_UNK_08000000;
-        entity->palette = 0x8195;
-        entity->animSet = ANIMSET_DRA(5);
-        entity->animCurFrame = 1;
-        entity->drawMode = DRAW_TPAGE;
-        entity->drawFlags = FLAG_DRAW_ROTX | FLAG_DRAW_ROTY;
-        temp_v0 = D_80180FBC[entity->params];
-        entity->rotX = temp_v0;
-        entity->rotY = temp_v0;
-        entity->velocityY = D_80180FCC[entity->params];
-        entity->step++;
-        return;
-    }
-
-    entity->animFrameDuration++;
-    entity->posY.val -= entity->velocityY;
-
-    if (!(entity->animFrameDuration & 1)) {
-        entity->animCurFrame++;
-    }
-
-    if (entity->animFrameDuration >= 37) {
-        DestroyEntity(entity);
-    }
-}
+#include "../entity_unkId15.h"
 
 #include "../entity_olrox_drool.h"
