+++ src/st/wrp_psp/wrp_data_F010.c
@@ -1,6 +1,6 @@
 #include "../wrp/wrp.h"
 
-u16 D_80180FF8[] = {
+u16 unk15_rot[] = {
     /* FF8 */ 0x0030,
     /* FFA */ 0x0050,
     /* FFC */ 0x0080,
@@ -11,7 +11,7 @@ u16 D_80180FF8[] = {
     /* 1006 */ 0x0000,
 };
 
-s32 D_80181008[] = {
+s32 unk15_yVel[] = {
     /* 1008 */ 0x00000400,
     /* 100C */ 0x00002400,
     /* 1010 */ 0x00003C00,
